Appointment service: Appointment and schedule services defined and primarily added.

+++ Appointment/src/main/java/com/healtcare/appointments/entities/Schedule.java
@@ -23,7 +23,7 @@ public class Schedule {
     private Integer afternoonShift;
     private Integer eveningShift;
 
-    private Long morningCapacity;
-    private Long afternoonCapacity;
+    private Integer morningCapacity;
+    private Integer afternoonCapacity;
     private Integer eveningCapacity;
 }
